deploy-flake: Replace rsync with nix flake archive
This changes the way to transfer the flake from the local to the remote
machine. This removes the dependency on `rsync` (and eventually `git` if
the flake has `git` inputs), and also private and local (e.g. for
development) inputs to be transferred without special setup.

This repurposes `lollypops.deployment.config-dir` to simply store the
link to the flake.

Potentially breaking change: The user needs to be able to write in the
Nix store, as I haven't figured a way to make nix flake archive use
sudo on the receiving side.