Remove `lollypops.deployment.config-dir`
This uses the fact that `self` (passed by `configFlake`) is a nix path,
and thus when transferred to a remote machine via `nix flake archive`
should result in the same path.

This is mostly for simplification, but also could open the path to
root-less building.

Breaking change: Something different happen in the case you run
deploy-flake, change something in the flake, and then run rebuild.
Before this commit, the system before modification would be built.
After this commit, the build will fail (as the flake path changed).